Awesome Updating Batman!
(Hopefully) the culmination of serviceworker updating notifications and also manual checking for updates "hidden" in the webapp changelog in settings.

Also on my custom android 9 ROM using Brave, the settings page's z-index looks lower than some of the main page so I fooled with that putting it after the webapp stuff and giving it a z-index as well.